[refactor] template스토리 파일 위치 변경
- 각 template내부 폴더로 이동
