+++ src/pages/Signup/template/SignupTemplate.stories.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import MockAdapter from 'axios-mock-adapter';
-import SignupTemplate from './template';
-import AxiosMock from '../../utils/AxiosMock';
-import { signupSuccessMockInfo, signupFailMockInfo } from '../../__mocks__/client-mock-data';
+import SignupTemplate from '.';
+import AxiosMock from '../../../utils/AxiosMock';
+import { signupSuccessMockInfo, signupFailMockInfo } from '../../../__mocks__/client-mock-data';
 
 export default {
   title: 'Template/Signup',
